# doc/apidoc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/news": {
      "get": {
        "operationId": "NewsController_list",
        "parameters": [
          {
            "name": "tournament",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "News"
        ]
      }
    },
    "/api/v1/news/{slug}": {
      "get": {
        "operationId": "NewsController_article",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "News"
        ]
      }
    },
    "/api/v1/orgs/{orgId}/news": {
      "post": {
        "operationId": "NewsController_create",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNewsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "News"
        ]
      }
    },
    "/api/v1/news-admin/{id}": {
      "patch": {
        "operationId": "NewsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNewsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "News"
        ]
      }
    },
    "/api/v1/news-admin/{id}/publish": {
      "post": {
        "operationId": "NewsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "News"
        ]
      }
    },
    "/api/v1/news-admin/{id}/unpublish": {
      "post": {
        "operationId": "NewsController_unpublish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "News"
        ]
      }
    },
    "/api/v1/cms/pages/{slug}": {
      "get": {
        "operationId": "CmsController_page",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CMS"
        ]
      }
    },
...
    "/api/v1/cms/admin/pages": {
      "get": {
        "operationId": "CmsController_pages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CMS"
        ]
      },
      "post": {
        "operationId": "CmsController_createPage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "CMS"
        ]
      }
    },
    "/api/v1/cms/admin/pages/{id}": {
      "get": {
        "operationId": "CmsController_pageDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CMS"
        ]
      },
      "patch": {
        "operationId": "CmsController_updatePage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CMS"
        ]
      }
    },
    "/api/v1/cms/admin/pages/{id}/publish": {
      "post": {
        "operationId": "CmsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "CMS"
        ]
      }
    },
    "/api/v1/cms/admin/pages/{id}/unpublish": {
      "post": {
        "operationId": "CmsController_unpublish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "CMS"
        ]
      }
    },
    "/api/v1/cms/admin/pages/{id}/revisions": {
      "get": {
        "operationId": "CmsController_revisions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CMS"
        ]
      }
    },
    "/api/v1/cms/admin/pages/{id}/revisions/{n}/restore": {
      "post": {
        "operationId": "CmsController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "n",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "CMS"
        ]
      }
    },
    "/api/v1/cms/admin/settings/{key}": {
      "put": {
        "operationId": "CmsController_setSetting",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CMS"
        ]
      }
    },
    "/api/v1/contact": {
      "post": {
        "operationId": "ContactController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Contact"
        ]
      }
    },
...
  },
  "info": {
    "title": "CricLive API",
    "description": "Cricket live scoring platform ? auth, organizations, tournaments, ball-by-ball scoring, stats, SaaS plans, CMS. Authorize with the access_token from POST /auth/login.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste access_token from /auth/login"
      }
    },
    "schemas": {
...
      "CreateNewsDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateNewsDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePageDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePageDto": {
        "type": "object",
        "properties": {}
      },
      "SettingDto": {
        "type": "object",
        "properties": {}
      },
      "ContactDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}